not recognizing PCB in resource.h
