+++ lakin.5/oss.c
@@ -17,7 +17,7 @@
 #include "shm.h"
 #include "oss.h"
 #include "queue.h"
-//#include "resource.h"
+#include "resource.h"
 
 /***
 	- allocate shm for data structures
@@ -47,7 +47,7 @@ int main(int argc, char ** argv){
         osclock.add(shm_data->launchSec, shm_data->launchNano);
     }
 
-    while (totalProcesses < 10) {
+    while (totalProcesses < 3) {
       scheduler();
     }
 
@@ -189,69 +189,13 @@ PCB * createProcess() {
     } else {
     }
 
-
     osclock.add(0,1);
     return pcb;
 }
 
 // ** Something not working right here, either first or last
-state * initializeResources() {
-    // initialize system resources
-    printf("\nin initialize\n");
-    int i;
-    for (i = 0; i < 20; i++) {
-      shm_data->r_state.resource[i] = shm_data->r_state.available[i] = rand() % MAX + 1;
-    }
-    // print system resources
-    printf("\nSystem Resources:\n");
-    for (i = 0; i < 20; i++) {
-        printf("R%02d ", i);
-    }
-    printf("\n");
-    for (i = 0; i < 20; i++) {
-         printf(" %02d ", shm_data->r_state.resource[i]);
-    }
-    // initialize available resources
-    // for (i = 0; i < 20; i++) {
-    //   shm_data->r_state.available[i] = shm_data->r_state.resource[i];
-    // }
-    printf("\nAvailable Resources:\n");
-    // print available resources
-    for (i = 0; i < 20; i++) {
-        printf("R%02d ", i);
-    }
-    printf("\n");
-    for (i = 0; i < 20; i++) {
-         printf(" %02d ", shm_data->r_state.available[i]);
-    }
-    printf("\n");
-}
 
-state * claimMatrix(PCB *pcb, int pcbIndex) {
-    int j;
 
-    for (j = 0; j < 20; j++){
-      shm_data->r_state.claim[pcbIndex][j] = pcb->rsrcsNeeded[j];
-
-    }
-}
-
-void printClaimMatrix() {
-    int i, j;
-    printf("\nClaim Matrix:\n");
-    printf("    ");
-    // print available resources
-    for (i = 0; i < 20; i++) {
-        printf("R%02d ", i);
-    }
-    for (i = 0; i < 18; i++) {
-      printf("\nP%02d ", i);
-      for(j = 0; j < 20; j++) {
-        printf(" %02d ", shm_data->r_state.claim[i][j]);
-      }
-    }
-    printf("\n");
-}
 
 /********** keep ************/
