+++ lakin.5/oss.h
@@ -21,8 +21,6 @@ void initialize();
 void initializeSharedMemory();
 void initializeMessageQueue();
 struct proc_ctrl_blck  * createProcess();
-struct state * initializeResources();
-struct state * claimMatrix(PCB*, int);
 void requestResponse();
 void createMessageQueue();
 void launchNewProc();
@@ -38,4 +36,3 @@ void sigHandler(const int);
 int initializeSig();
 void scheduler();
 int findAvailablePcb(void);
-void printClaimMatrix();
