+++ lakin.5/resource.h
@@ -1,1 +1,5 @@
+#pragma once
 
+struct state * initializeResources();
+struct state * claimMatrix(PCB *, int);
+void printClaimMatrix();
